Implemented screen namespace and fixed 0xD instruction
The 0xD instruction was always using the memory address in the index
register to draw all the rows. It turns out a new byte needs to be read
per row in order to draw the sprite properly.
The resulting display can now be seen and there is a window for the
emulator.